app/requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
markdown==3.5.1
reportlab==4.0.7